Add lens component, move MD from DC to lens
To reduce the the depthOfField, which makes the focus blur stronger, adjust the magnification, refractive index and numerical aperture on the lens component. The metadata updater takes care that the values on those VAs on the lens are put onto the metadata on the diagnostic camera. It also calculates then the MD_PIXEL_SIZE as sensor px size / magnification * binning.
Don't put the info as MD directly on the DC.
The pixel size is needed to calculate the depth of field.
The refractive index needs to be equal greater than 1.
Update test cases to use stage instead of diagnostic-cam-focus.